06_week/project/etl/app/requirements.txt
pymongo==3.11.3
psycopg2-binary==2.8.6
SQLAlchemy==1.4.11
vaderSentiment==3.3.2